Add missing stub classes to news_trading
