+++ backend/alert_system.py
@@ -1,0 +1,5 @@
+class AlertSystem:
+    """Placeholder alert system."""
+
+    def send_alert(self, *args, **kwargs):
+        pass
