+++ backend/central_logic_engine.py
@@ -1,0 +1,5 @@
+class CentralLogicEngine:
+    """Minimal central logic engine for tests."""
+
+    def __init__(self, **components):
+        self.components = components
