+++ backend/enhanced_backtesting.py
@@ -1,0 +1,5 @@
+class EnhancedBacktester:
+    """Placeholder enhanced backtester used for tests."""
+
+    def run_backtest(self, *args, **kwargs):
+        return {"trades": [], "performance_metrics": {}, "equity_curve": []}
