+++ backend/gemma3_integration/natural_language_market_analysis_enhanced.py
@@ -889,3 +889,7 @@ class NaturalLanguageMarketAnalyzer:
             }
         
         return analysis
+
+# Alias for backward compatibility
+NaturalLanguageMarketAnalysis = NaturalLanguageMarketAnalyzer
+
