+++ backend/paper_trading_mode.py
@@ -1,0 +1,17 @@
+class PaperTradingEngine:
+    """Placeholder paper trading engine."""
+
+    def __init__(self, **kwargs):
+        pass
+
+    def initialize(self):
+        pass
+
+    def execute_trade(self, *args, **kwargs):
+        pass
+
+    def get_portfolio_status(self):
+        return {}
+
+    def get_trade_history(self):
+        return []
