+++ backend/portfolio_optimization.py
@@ -1,0 +1,5 @@
+class PortfolioOptimizer:
+    """Placeholder portfolio optimizer."""
+
+    def optimize(self, *args, **kwargs):
+        return {}
