+++ requirements_essential.txt
@@ -34,3 +34,6 @@ pytest==7.4.0
 tqdm==4.66.3
 python-dotenv==1.0.0
 joblib==1.3.2
+nltk==3.9.1
+psutil==5.9.8
+textblob==0.18.0
